Fix, SIT-510, SIT-479, SIT-480. Confirmation dialogs when abortin/turning an exam and canceling reservation

+++ public/javascript/controllers/dashboardController.js
@@ -91,10 +91,13 @@
                 }
 
                 $scope.removeReservation = function(enrolment){
-                    $http.delete('calendar/reservation/' +enrolment.reservation.id).success(function(){
-                        delete enrolment.reservation;
-                        toastr.success("ok");
-                    });
+                    if (confirm($translate('sitnet_are_you_sure'))) {
+
+                        $http.delete('calendar/reservation/' + enrolment.reservation.id).success(function () {
+                            delete enrolment.reservation;
+                            toastr.success("ok");
+                        });
+                    }
                 };
 
                 $scope.beforeDate = function(date) {
